Feature invoice

+++ database/migrations/2019_11_18_172345_create_companies_table.php
@@ -16,9 +16,9 @@ class CreateCompaniesTable extends Migration
         Schema::create('companies', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->string('name');
-            $table->integer('nit')->unique();
+            $table->unsignedBigInteger('nit')->unique();
             $table->string('email')->nullable();
-            $table->integer('phone')->nullable();
+            $table->unsignedBigInteger('phone')->nullable();
             $table->string('address')->nullable();
             $table->timestamps();
         });
